# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Semantic engine
  version: "0.1.0"
  description: |
    An Apache Ossie semantic model compiled to governed, dialect-correct SQL.

    There is no endpoint that accepts SQL, and there will not be one. The only
    expressible request is a semantic one: metrics, dimensions and structured
    filters. That is what makes an ungoverned query inexpressible rather than
    merely discouraged.

    Column access is resolved before SQL is emitted. A denied request returns a
    refusal with a reason; it never returns an empty result set, because a
    silent empty answer trains a caller to report a wrong number confidently.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://127.0.0.1:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /v1/audit:
    get:
      tags: [governance]
      operationId: listAudit
      summary: Read the decisions this engine recently made
      description: |
        Every decision the engine made, newest first: what was allowed, what
        was refused for correctness, what was denied for access, and what
        failed. This is the window an operator reads to answer "why did that
        agent give up", which nothing else can answer.

        Not served unless an operator names who may read it. The record
        contains every caller's activity, which discloses what other teams
        query and which fields are protected, so an engine with no reader
        configured answers 404: the capability is absent rather than withheld.
        A caller who is authenticated but not a named reader gets 403.

        This is a bounded in-memory window, not the whole record. An engine
        started with an audit file keeps everything there; this is what it can
        still show you without shelling into the container.

        Carries no filter values, no compiled SQL and no result rows, for the
        same reason the audit file does not.
      parameters:
        - name: limit
          in: query
          required: false
          description: Most recent events to return.
          schema: { type: integer, minimum: 1, maximum: 200, default: 200 }
        - name: decision
          in: query
          required: false
          description: Return only this decision. `refused` is the usual one.
          schema:
            type: string
            enum: [allowed, refused, denied, compiled, error]
      responses:
        "200":
          description: Recorded decisions, newest first
          content:
            application/json:
              schema:
                type: object
                required: [events, count]
                properties:
                  events:
                    type: array
                    items: { $ref: "#/components/schemas/AuditEvent" }
                  count: { type: integer }
                  note:
                    type: string
                    description: Says that this is a window rather than the whole record.
        "400": { $ref: "#/components/responses/Refused" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Refused" }
        "404": { $ref: "#/components/responses/Refused" }
...
  /v1/tests:
    post:
      tags: [governance]
      operationId: runTests
      summary: Assert what this model answers
      description: |
        validate says the model holds together and diff says a number changed.
        Neither says a number was ever right, which is the gap an adopter
        notices first because correctness is the entire pitch.

        Refusal assertions need no warehouse, so a caller holding only
        read:model can still prove the model declines a fan-out. Cases that
        would execute a query are withheld from such a caller and counted,
        rather than run or silently dropped: a suite that quietly checked half
        of what it claimed is worse than one that failed.

        404 until an operator names a test suite. An engine nobody pointed at
        one has no tests, which is a different thing from having tests that
        all pass.
      responses:
        "200":
          description: The result of every case
          content:
            application/json:
              schema:
                type: object
                required: [ok, passed, failed, skipped, results]
                properties:
                  ok: { type: boolean }
                  passed: { type: integer }
                  failed: { type: integer }
                  skipped: { type: integer }
                  withheld:
                    type: integer
                    description: Cases this credential may not run.
                  results:
                    type: array
                    items: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Refused" }
        "404": { $ref: "#/components/responses/Refused" }
        "422": { $ref: "#/components/responses/Refused" }
...
  /v1/policy/explain:
    post:
      tags: [governance]
      operationId: explainPolicy
      summary: What you may read of a metric
      description: |
        Answers why a column cannot be grouped by, without running a query and
        being denied.

        For the calling identity only, and that is a security property rather
        than a limitation. An endpoint that reported what somebody else can
        see would publish the policy it was configured to enforce, so asking
        about another identity is a different endpoint with its own
        authorization and this engine does not serve one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [metric]
              properties:
                metric:
                  type: string
                  description: Qualified metric name, as /v1/metrics reports it.
      responses:
        "200":
          description: The dimensions this caller may group that metric by
          content:
            application/json:
              schema:
                type: object
                required: [metric, readable]
                properties:
                  metric: { type: string }
                  identity: { type: string }
                  readable:
                    type: array
                    items: { type: string }
                  governance: { type: object }
        "400": { $ref: "#/components/responses/Refused" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Refused" }
        "404": { $ref: "#/components/responses/Refused" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        The token identifies a workload, and that workload identity is what the
        governance gate and the executor use. Grants go to service accounts and
        groups, never to individual humans.
  responses:
    Unauthorized:
      description: Missing or unrecognised credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Refusal" }
    Refused:
      description: |
        The engine understood the request and declined it. Branch on `retry`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Refusal" }
  schemas:
    AuditEvent:
      type: object
      description: |
        One decision, as recorded. `refused` and `denied` are deliberately
        separate: denied is access, meaning this caller may not read something,
        and refused is correctness, meaning nobody can be told this accurately
        and the hint names a question that can be answered. Reporting them
        together would count every fan-out as an access incident.
      required: [time, decision]
      properties:
        time: { type: string, format: date-time }
        identity:
          type: string
          description: Who asked. `anonymous` on an engine without authentication.
        model_name: { type: string }
        model_version: { type: string }
        namespace: { type: string }
        metrics:
          type: array
          items: { type: string }
        dimensions:
          type: array
          items: { type: string }
        decision:
          type: string
          enum: [allowed, refused, denied, compiled, error]
        refusal_code:
          type: string
          description: Which refusal, for a decision of `refused`.
        retry:
          type: string
          enum: [modify, later, never]
          description: What the caller should do about a refusal.
        reason:
          type: string
          description: The refusal in the engine's own words, as the caller received it.
        hint:
          type: string
          description: What would answer instead. The actionable half of a refusal.
        denied_fields:
          type: array
          description: Semantic fields refused, on a denial. Never returned to the denied caller.
          items: { type: string }
        sql_hash:
          type: string
          description: Identifies the compiled statement without disclosing it.
        dialect: { type: string }
        job_id:
          type: string
          description: The warehouse's own identifier, so two logs can be joined.
        row_count: { type: integer }
        bytes_billed:
          type: integer
          format: int64
          description: |
            What the warehouse says the query cost. Absent means not reported
            rather than free: DuckDB bills nobody and reports nothing.
        duration_ms: { type: integer, format: int64 }
        error:
          type: string
          description: The failure, for a decision of `error`.
    Refusal:
      type: object
      required: [code, reason, retry]
      properties:
        code:
          type: string
          description: Machine-readable reason. Branch on this, not on the text.
          examples: [fan_out_would_inflate, access_denied, unknown_metric]
        reason:
          type: string
          description: One sentence stating what was wrong.
        hint:
          type: string
          description: |
            What to do instead. For a fan-out refusal this names the metrics
            defined at the grain where the question is well defined.
        retry: { $ref: "#/components/schemas/Retry" }
    Retry:
      type: string
      enum: [modify, later, never]
      description: |
        What a caller should do about a refusal. This is the field an agent
        branches on; without it a caller either gives up on a typo or loops
        forever on a denial.

        `modify`: the request is answerable but not as written. Change the
        arguments and try again; repeating it unchanged will not work.

        `later`: something outside the request failed, such as the policy
        source being unreachable, or the caller is at their concurrency limit.
        The same request may succeed later.

        `never`: no version of this request from this caller will succeed.
        Stop, and say so rather than guessing.
...